PIPELINE: move test files - both test files are moved to 'test' directory - replace os.popen with subprocess.check_output instead 	- add new cwd for subprocess
